refactor(directory) Move auth to middleware package
