+++ pkg/middleware/auth.go
@@ -1,4 +1,4 @@
-package api
+package middleware
 
 import (
 	"context"
